refactor: core:database -> local:common
