--- core/database/src/main/java/com/suwiki/core/database/dao/TimeTableDao.kt
@@ -1,29 +1,0 @@
-package com.suwiki.core.database.dao
-
-import androidx.room.Dao
-import androidx.room.Delete
-import androidx.room.Insert
-import androidx.room.Query
-import androidx.room.Update
-import com.suwiki.core.database.model.TimetableEntity
-
-@Dao
-interface TimeTableDao {
-  @Query("SELECT * FROM TimetableEntity ORDER BY createTime DESC")
-  fun getAll(): List<TimetableEntity>
-
-  @Query("SELECT * FROM TimetableEntity WHERE createTime = :createTime")
-  fun get(createTime: Long): TimetableEntity?
-
-  @Insert
-  fun insert(data: TimetableEntity)
-
-  @Query("DELETE FROM TimetableEntity")
-  fun deleteAll()
-
-  @Delete
-  fun delete(data: TimetableEntity)
-
-  @Update
-  fun update(data: TimetableEntity)
-}
