--- core/database/src/main/java/com/suwiki/core/database/di/NormalDataStore.kt
@@ -1,7 +1,0 @@
-package com.suwiki.core.database.di
-
-import javax.inject.Qualifier
-
-@Qualifier
-@Retention(AnnotationRetention.BINARY)
-annotation class NormalDataStore
