--- core/database/src/main/java/com/suwiki/core/database/model/OpenMajorEntity.kt
@@ -1,12 +1,0 @@
-package com.suwiki.core.database.model
-
-import androidx.room.Entity
-import androidx.room.PrimaryKey
-
-@Entity
-data class OpenMajorEntity(
-  val name: String,
-) {
-  @PrimaryKey(autoGenerate = true)
-  var id: Int = 0
-}
