--- core/database/src/main/java/com/suwiki/core/database/model/TimetableEntity.kt
@@ -1,14 +1,0 @@
-package com.suwiki.core.database.model
-
-import androidx.room.Entity
-import androidx.room.PrimaryKey
-import com.suwiki.core.model.timetable.TimetableCell
-
-@Entity
-data class TimetableEntity(
-  @PrimaryKey val createTime: Long,
-  var year: String,
-  var semester: String,
-  var name: String,
-  var cellList: List<TimetableCell>,
-)
