--- core/database/src/main/java/com/suwiki/core/database/proto/UserPreferenceSerializer.kt
@@ -1,24 +1,0 @@
-package com.suwiki.core.database.proto
-
-import androidx.datastore.core.CorruptionException
-import androidx.datastore.core.Serializer
-import androidx.datastore.preferences.protobuf.InvalidProtocolBufferException
-import com.suwiki.core.database.UserPreference
-import java.io.InputStream
-import java.io.OutputStream
-import javax.inject.Inject
-
-class UserPreferenceSerializer @Inject constructor() : Serializer<UserPreference> {
-  override val defaultValue: UserPreference = UserPreference.getDefaultInstance()
-
-  override suspend fun readFrom(input: InputStream): UserPreference =
-    try {
-      UserPreference.parseFrom(input)
-    } catch (exception: InvalidProtocolBufferException) {
-      throw CorruptionException("Cannot read proto.", exception)
-    }
-
-  override suspend fun writeTo(t: UserPreference, output: OutputStream) {
-    t.writeTo(output)
-  }
-}
