--- core/database/src/main/proto/user_preference.proto
@@ -1,13 +1,0 @@
-syntax = "proto3";
-
-option java_package = "com.suwiki.core.database";
-option java_multiple_files = true;
-
-message UserPreference {
-  string user_id = 1;
-  string email = 2;
-  int32 point = 3;
-  int32 written_evaluation = 4;
-  int32 written_exam = 5;
-  int32 view_exam = 6;
-}
