+++ local/common/src/main/java/com/suwiki/local/common/database/entity/OpenMajorEntity.kt
@@ -1,0 +1,12 @@
+package com.suwiki.local.common.database.entity
+
+import androidx.room.Entity
+import androidx.room.PrimaryKey
+
+@Entity
+data class OpenMajorEntity(
+  val name: String,
+) {
+  @PrimaryKey(autoGenerate = true)
+  var id: Int = 0
+}
