+++ local/common/src/main/java/com/suwiki/local/common/database/entity/TimetableEntity.kt
@@ -1,0 +1,14 @@
+package com.suwiki.local.common.database.entity
+
+import androidx.room.Entity
+import androidx.room.PrimaryKey
+import com.suwiki.core.model.timetable.TimetableCell
+
+@Entity
+data class TimetableEntity(
+  @PrimaryKey val createTime: Long,
+  var year: String,
+  var semester: String,
+  var name: String,
+  var cellList: List<TimetableCell>,
+)
