+++ local/common/src/main/java/com/suwiki/local/common/datastore/di/NormalDataStore.kt
@@ -1,0 +1,7 @@
+package com.suwiki.local.common.datastore.di
+
+import javax.inject.Qualifier
+
+@Qualifier
+@Retention(AnnotationRetention.BINARY)
+annotation class NormalDataStore
