+++ local/timetable/build.gradle.kts
@@ -13,7 +13,7 @@ android {
 dependencies {
   implementation(projects.core.model)
   implementation(projects.core.android)
-  implementation(projects.core.database)
+  implementation(projects.local.common)
 
   implementation(projects.data.timetable)
