+++ local/timetable/src/main/java/com.suwiki.local.timetable/converter/TimetableEntity.kt
@@ -1,7 +1,7 @@
 package com.suwiki.local.timetable.converter
 
-import com.suwiki.core.database.model.TimetableEntity
 import com.suwiki.core.model.timetable.Timetable
+import com.suwiki.local.common.database.entity.TimetableEntity
 
 fun TimetableEntity.toModel() = Timetable(
   createTime = createTime,
