+++ local/user/build.gradle.kts
@@ -11,7 +11,7 @@ android {
 dependencies {
   implementation(projects.core.model)
   implementation(projects.data.user)
-  implementation(projects.core.database)
+  implementation(projects.local.common)
 
   implementation(libs.bundles.coroutine)
   implementation(libs.androidx.datastore.core)
