--- schemas/com.suwiki.core.database.database.OpenLectureDatabase/2.json
@@ -1,88 +1,0 @@
-{
-  "formatVersion": 1,
-  "database": {
-    "version": 2,
-    "identityHash": "d388602654d094c34e2c718a802d0243",
-    "entities": [
-      {
-        "tableName": "OpenLectureEntity",
-        "createSql": "CREATE TABLE IF NOT EXISTS `${TABLE_NAME}` (`number` INTEGER NOT NULL, `major` TEXT NOT NULL, `grade` TEXT NOT NULL, `classNumber` TEXT NOT NULL, `classDivideNumber` TEXT NOT NULL, `className` TEXT NOT NULL, `classification` TEXT NOT NULL, `professor` TEXT NOT NULL, `time` TEXT NOT NULL, `credit` TEXT NOT NULL, PRIMARY KEY(`number`))",
-        "fields": [
-          {
-            "fieldPath": "number",
-            "columnName": "number",
-            "affinity": "INTEGER",
-            "notNull": true
-          },
-          {
-            "fieldPath": "major",
-            "columnName": "major",
-            "affinity": "TEXT",
-            "notNull": true
-          },
-          {
-            "fieldPath": "grade",
-            "columnName": "grade",
-            "affinity": "TEXT",
-            "notNull": true
-          },
-          {
-            "fieldPath": "classNumber",
-            "columnName": "classNumber",
-            "affinity": "TEXT",
-            "notNull": true
-          },
-          {
-            "fieldPath": "classDivideNumber",
-            "columnName": "classDivideNumber",
-            "affinity": "TEXT",
-            "notNull": true
-          },
-          {
-            "fieldPath": "className",
-            "columnName": "className",
-            "affinity": "TEXT",
-            "notNull": true
-          },
-          {
-            "fieldPath": "classification",
-            "columnName": "classification",
-            "affinity": "TEXT",
-            "notNull": true
-          },
-          {
-            "fieldPath": "professor",
-            "columnName": "professor",
-            "affinity": "TEXT",
-            "notNull": true
-          },
-          {
-            "fieldPath": "time",
-            "columnName": "time",
-            "affinity": "TEXT",
-            "notNull": true
-          },
-          {
-            "fieldPath": "credit",
-            "columnName": "credit",
-            "affinity": "TEXT",
-            "notNull": true
-          }
-        ],
-        "primaryKey": {
-          "autoGenerate": false,
-          "columnNames": [
-            "number"
-          ]
-        },
-        "indices": [],
-        "foreignKeys": []
-      }
-    ],
-    "views": [],
-    "setupQueries": [
-      "CREATE TABLE IF NOT EXISTS room_master_table (id INTEGER PRIMARY KEY,identity_hash TEXT)",
-      "INSERT OR REPLACE INTO room_master_table (id,identity_hash) VALUES(42, 'd388602654d094c34e2c718a802d0243')"
-    ]
-  }
-}
