--- schemas/com.suwiki.core.database.database.TimetableDatabase/2.json
@@ -1,58 +1,0 @@
-{
-  "formatVersion": 1,
-  "database": {
-    "version": 2,
-    "identityHash": "d3ae462e5e5245e675c522c8e1ff9800",
-    "entities": [
-      {
-        "tableName": "TimetableEntity",
-        "createSql": "CREATE TABLE IF NOT EXISTS `${TABLE_NAME}` (`createTime` INTEGER NOT NULL, `year` TEXT NOT NULL, `semester` TEXT NOT NULL, `name` TEXT NOT NULL, `cellList` TEXT NOT NULL, PRIMARY KEY(`createTime`))",
-        "fields": [
-          {
-            "fieldPath": "createTime",
-            "columnName": "createTime",
-            "affinity": "INTEGER",
-            "notNull": true
-          },
-          {
-            "fieldPath": "year",
-            "columnName": "year",
-            "affinity": "TEXT",
-            "notNull": true
-          },
-          {
-            "fieldPath": "semester",
-            "columnName": "semester",
-            "affinity": "TEXT",
-            "notNull": true
-          },
-          {
-            "fieldPath": "name",
-            "columnName": "name",
-            "affinity": "TEXT",
-            "notNull": true
-          },
-          {
-            "fieldPath": "cellList",
-            "columnName": "cellList",
-            "affinity": "TEXT",
-            "notNull": true
-          }
-        ],
-        "primaryKey": {
-          "autoGenerate": false,
-          "columnNames": [
-            "createTime"
-          ]
-        },
-        "indices": [],
-        "foreignKeys": []
-      }
-    ],
-    "views": [],
-    "setupQueries": [
-      "CREATE TABLE IF NOT EXISTS room_master_table (id INTEGER PRIMARY KEY,identity_hash TEXT)",
-      "INSERT OR REPLACE INTO room_master_table (id,identity_hash) VALUES(42, 'd3ae462e5e5245e675c522c8e1ff9800')"
-    ]
-  }
-}
