+++ schemas/com.suwiki.local.common.database.TimetableDatabase/1.json
