+++ settings.gradle.kts
@@ -26,7 +26,6 @@ include(":core:model")
 include(":core:common")
 include(":core:network")
 include(":core:security")
-include(":core:database")
 include(":core:designsystem")
 include(":core:ui")
 
@@ -75,3 +74,4 @@ include(":feature:myinfo")
 include(":feature:notice")
 include(":feature:signup")
 include(":feature:login")
+include(":local:common")
